pyproject.toml
[project]
name = "rag-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.73.1",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langchain-openai>=1.1.14",
    "langchain-text-splitters>=1.1.2",
    "python-dotenv>=1.2.2",
]
